refactor: split router.py into 10 cohesive sub-modules (iter29)
Decompose the router.py god-file (2628 raw / ~1600 NLOC -> 488 NLOC) into 10 sibling
modules each <500 NLOC: router_conn, router_directplay, router_dispatch,
router_episodeinfo, router_fallback, router_play, router_poster, router_scriptplay,
router_search, router_settings.

Proven net-positive pattern: every test-@patched name (route, the _handle_* entry
points, _search_all_providers, _test_*, etc.) stays in router.py or is re-exported via
__all__; moved helpers reach router-resident/patched names via function-local
`import resources.lib.router as _router` so @patch resolves at call time and no
top-level cycle forms; no dependencies threaded as params (params>8 flat at 16).
setResolvedUrl/endOfDirectory/loading-dialog ordering preserved.

Codacy complexity 27 -> 26 (file-nloc>500: 4->3). Verified: 2138/2140 tests pass
(2 pre-existing wall-clock timing flakes, unrelated - test_resolver doesn't import
router), pylint 10/10, ruff+black clean, vermin 3.8 OK.
